Move interpolation types to a separate file
