+++ src/i2pp/core/interpolators/interpolator_types.py
@@ -1,0 +1,60 @@
+"""Interpolation type definitions and handling."""
+
+from enum import Enum
+from typing import Type
+
+from i2pp.core.interpolators.interpolator import Interpolator
+from i2pp.core.interpolators.interpolator_all_voxel import InterpolatorAllVoxel
+from i2pp.core.interpolators.interpolator_center import InterpolatorCenter
+from i2pp.core.interpolators.interpolator_nodes import InterpolatorNodes
+
+
+class InterpolationType(Enum):
+    """Enum representing different calculation types for element value
+    determination.
+
+    This enum defines the available calculation types for mapping image data
+    to finite element Discretization elements. The calculation type determines
+    how the image pixel data is assigned to the elements in the
+    Discretization.
+
+    Attributes:
+        NODES (str): Represents the calculation method where the pixel value
+            is averaged over the nodes of the element.
+        CENTER (str): Represents the calculation method where the pixel value
+            is based on the center of the element.
+        ALLVOXELS (str): Represents the calculation method where the pixel
+            value is averaged over all voxels inside the element.
+    """
+
+    NODES = "nodes"
+    CENTER = "elementcenter"
+    ALLVOXELS = "allvoxels"
+
+    def get_interpolator(self) -> Type[Interpolator]:
+        """Retrieves the appropriate interpolation class based on the selected
+        calculation type.
+
+        This method returns the corresponding interpolator class for assigning
+        pixel values to FEM elements, depending on the current calculation
+        type. Supported methods include interpolation at element nodes, element
+        centers, or averaging all voxels within an element.
+
+        Returns:
+            Type[Interpolator]: The interpolator class that matches the
+                specified calculation type.
+
+        Raises:
+            ValueError: If the calculation type is not supported.
+        """
+
+        interpolator_map = {
+            InterpolationType.NODES: InterpolatorNodes,
+            InterpolationType.ALLVOXELS: InterpolatorAllVoxel,
+            InterpolationType.CENTER: InterpolatorCenter,
+        }
+
+        if self not in interpolator_map:
+            raise ValueError(f"Unsupported interpolation method: {self}")
+
+        return interpolator_map[self]
